move memory location test
